Early tutorial + Proximity GUI Change
Early implementation of tutorials using text with delay to show up and disappear.
Proximity GUI is no more. For now. In the future it may be used as a diegetic form of letting the player know how distant it is from the nearby planet. For now it calls are commented.